requirements.txt
Django==2.2
pytz==2019.1
sqlparse==0.3.0
django-cors-headers==2.5.2
django-crontab==0.7.1
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-stubs==0.2.0
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7